@@ -1,4 +1,4 @@
-﻿@model BlogEngine.DataTransferObject.Profile
+﻿@model BlogEngine.DataTransferObject.UserProfile
 
 @{
     ViewData["Title"] = "About";
